Refactor manifest tasks
