+++ automation_tools/__init__.py
@@ -4,13 +4,11 @@ Many commands are affected by environment variables. Unless stated otherwise,
 all environment variables are required.
 """
 from __future__ import print_function
-import base64
 import os
 import random
 import socket
 import sys
 import time
-from datetime import date
 from io import StringIO
 from re import MULTILINE, search
 
@@ -2942,223 +2940,6 @@ def setenforce(mode):
         ))
 
     run('setenforce {0}'.format(mode))
-
-
-def download_manifest(url, consumer, user, password):
-    """Task for downloading the manifest file from customer portal.
-
-    :param url: Subscription Manager URL
-    :param consumer: A consumer hash to be used for getting the manifest
-    :param user: Red Hat Network username
-    :param password: Red Hat Network password
-    :returns: a path string to a downloaded manifest file
-    """
-    string = u'{0}:{1}'.format(user, password)
-
-    if isinstance(string, str):  # py3
-        bytestring = bytes('{0}:{1}'.format(user, password), 'utf-8')
-    else:  # py2
-        bytestring = bytes('{0}:{1}'.format(user, password))
-
-    base64string = base64.encodestring(bytestring).strip()
-
-    manifest_file = run('mktemp --suffix=.zip')
-
-    # we do this as we would otherwise potentially download a manifest which
-    # has new metadata (such as new content sets/repos) but does not have the
-    # required Entitlement certificates to actually access the content.
-    certs_put = ('curl -sk -X PUT -H "Authorization:Basic {0}"'
-                 ' {1}/subscription/consumers/{2}/certificates'
-                 '?lazy_regen=false').format(
-                 base64string.decode('utf-8'), url, consumer)
-    run(certs_put)
-
-    command = ('curl -sk -H "Authorization:Basic {0}"'
-               ' {1}/subscription/consumers/{2}/export/').format(
-               base64string.decode('utf-8'), url, consumer)
-
-    response = run(command + ' -I')
-    if ('Content-Disposition: attachment' in response):
-        run(command + ' -o {0}'.format(manifest_file))
-        return manifest_file
-    else:
-        raise ValueError('Request has returned no attachment. Check the'
-                         ' session and distributor hash')
-
-
-def validate_manifest(user, password, url, consumer, manifest_file, exp_subs_file):
-    """Make sure that manifest contains only subscriptions specified in config file
-    specified in variable and attach if missing any:
-
-    :param user: Red Hat Network username
-    :param password: Red Hat Network password
-    :param exp_subs_file: Expected subscription file
-    :param manifest_file: Specify the manifest file path.
-    :return: list with True/False boolean
-    """
-    with open(exp_subs_file, 'r') as fp:
-        exp_subs = {}
-        exp_subs_details = [r.strip().split(';') for r in fp.readlines() if not r.strip(
-        ).startswith('#')]
-        for name_list in exp_subs_details:
-            exp_subs[name_list[0]] = name_list[1:]
-
-    print("expected subscriptions in {} are {}".format(manifest_file, ' & '.join(exp_subs.keys())))
-    rct_output = run("rct cat-manifest --no-content {}".format(manifest_file)).split("\n")
-    sub_name_next = False
-    current_subs = list()
-    for r in rct_output:
-        r = r.strip()
-        if r == 'Subscription:':
-            sub_name_next = True
-            continue
-        if sub_name_next:
-            if r.startswith('Name: '):
-                current_subs.append(r[6:])
-            sub_name_next = False
-    print("current subscriptions in {} are {}".format(manifest_file, ' & '.join(current_subs)))
-
-    response = [True]
-    for sub in exp_subs.keys():
-        if sub not in current_subs:
-            response.append('Yes')
-            print("Attaching missing subscription {}".format(sub))
-            attach = attach_subscription(user=user, password=password,
-                                         url=url, consumer=consumer,
-                                         pool_id=exp_subs[sub][0],
-                                         count=exp_subs[sub][1])
-            if attach:
-                print("Successfully attached subscription {}" .format(sub))
-            else:
-                return [False]
-    return response
-
-
-def attach_subscription(user, password, url, consumer, pool_id=None, count=None):
-    """Task to attach subscription to manifest.
-
-    :param url: Subscription Manager URL
-    :param consumer: A consumer hash to be used for getting the manifest
-    :param user: Red Hat Network username
-    :param password: Red Hat Network password
-    :param pool_id: A pool ID of subscription
-    :param count: A quantity of subscription requires to attach
-    :returns: boolean True/False
-    """
-    auth_details = u'{0}:{1}'.format(user, password)
-
-    if isinstance(auth_details, str):  # py3
-        bytestring = bytes('{0}:{1}'.format(user, password), 'utf-8')
-    else:  # py2
-        bytestring = bytes('{0}:{1}'.format(user, password))
-
-    base64string = base64.encodestring(bytestring).strip()
-
-    if url is None:
-        url = os.environ.get('SM_URL')
-    if consumer is None:
-        consumer = os.environ.get('CONSUMER')
-
-    if pool_id and count:
-        command = ('curl -sk -X POST -H "Authorization:Basic {0}"'
-                   ' "{1}/subscription/consumers/{2}/entitlements?pool={3}&quantity={4}"').format(
-            base64string.decode('utf-8'), url, consumer, pool_id, count)
-        response = run(command)
-        return 'updated' in response
-    else:
-        raise ValueError("Pass values under pool_id and count")
-
-
-def refresh_manifest(url, consumer, user, password, exp_subs_file):
-    """Task to remove all subscriptions from manifest and attach required ones.
-
-    :param url: Subscription Manager URL
-    :param consumer: A consumer hash to be used for getting the manifest
-    :param user: Red Hat Network username
-    :param password: Red Hat Network password
-    :param exp_subs_file: Expected subscription file
-    :returns: boolean True/False
-    """
-    auth_details = u'{0}:{1}'.format(user, password)
-
-    if isinstance(auth_details, str):  # py3
-        bytestring = bytes('{0}:{1}'.format(user, password), 'utf-8')
-    else:  # py2
-        bytestring = bytes('{0}:{1}'.format(user, password))
-
-    base64string = base64.encodestring(bytestring).strip()
-
-    command = ('curl -sk -X DELETE -H "Authorization:Basic {0}"'
-               ' "{1}/subscription/consumers/{2}/entitlements"').format(
-        base64string.decode('utf-8'), url, consumer)
-
-    print("Cleaning up all subscriptions from Manifest!")
-    response = run(command)
-    result = 'deletedRecords' in response
-
-    if result:
-        with open(exp_subs_file, 'r') as fp:
-            exp_subs = {}
-            exp_subs_details = [r.strip().split(';') for r in fp.readlines() if not r.strip(
-            ).startswith('#')]
-            for name_list in exp_subs_details:
-                exp_subs[name_list[0]] = name_list[1:]
-
-            for sub in exp_subs.keys():
-                print("Attaching subscription {}".format(sub))
-                attach = attach_subscription(user=user, password=password,
-                                             url=url, consumer=consumer,
-                                             pool_id=exp_subs[sub][0],
-                                             count=exp_subs[sub][1])
-                if attach:
-                    print("Successfully attached subscription {}".format(sub))
-                else:
-                    return attach
-    return result
-
-
-def relink_manifest(url, consumer, user, password, exp_subs_file,
-                    manifest_file=None):
-    """Links the latest downloaded manifest file to the manifest_latest.zip
-    softlink.
-
-    :param url: Subscription Manager URL
-    :param consumer: A consumer hash to be used for getting the manifest
-    :param user: Red Hat Network username
-    :param password: Red Hat Network password
-    :param exp_subs_file: Expected subscription file
-    :param manifest_file: Specify the manifest file path.
-    """
-    if manifest_file is None:
-        manifest_file = download_manifest(url=url, consumer=consumer,
-                                          user=user, password=password)
-        if exp_subs_file is not None:
-            assert refresh_manifest(url=url, consumer=consumer,
-                                    user=user, password=password,
-                                    exp_subs_file=exp_subs_file)
-            validate = validate_manifest(user=user, password=password,
-                                         url=url, consumer=consumer,
-                                         manifest_file=manifest_file,
-                                         exp_subs_file=exp_subs_file)
-            assert validate[0]
-            if len(validate) > 1:
-                manifest_file = download_manifest(url=url, consumer=consumer,
-                                                  user=user, password=password)
-                assert validate_manifest(user=user, password=password,
-                                         url=url, consumer=consumer,
-                                         manifest_file=manifest_file,
-                                         exp_subs_file=exp_subs_file)[0]
-    if not manifest_file:
-        print('manifest_file is not populated.')
-        sys.exit(1)
-    date_str = date.today().strftime("%Y%m%d")
-    new_manifest_file = '/opt/manifests/manifest-{0}.zip'.format(date_str)
-    run('mv {0} {1}'.format(manifest_file, new_manifest_file))
-    run('chmod 644 {0}'.format(new_manifest_file))
-    run('restorecon -v {0}'.format(new_manifest_file))
-    run('unlink /opt/manifests/manifest-latest.zip')
-    run('ln -s {0} /opt/manifests/manifest-latest.zip'
-        .format(new_manifest_file))
 
 
 def set_yum_debug_level(level=1):
